refactor: rename file to match the exported function
