+++ src/features/search/atoms/Hit.tsx
@@ -4,7 +4,7 @@ import { useQueryClient } from 'react-query'
 import styled from 'styled-components/native'
 
 import { mergeOfferData } from 'features/offer/atoms/OfferTile'
-import { useLogClickOnOffer } from 'libs/algolia/analytics/logClickOnProduct'
+import { useLogClickOnOffer } from 'libs/algolia/analytics/logClickOnOffer'
 import { analytics } from 'libs/analytics'
 import { useDistance } from 'libs/geolocation/hooks/useDistance'
 import { formatDates, getDisplayPrice } from 'libs/parsers'
